200212_kijoon_결재 마무리, 호텔정보 마무리, 수리현황 추가

See merge request ghks8198/hotelscompile!118

+++ src/main/java/com/kh/hotels/mngApproval/model/vo/RepRequest.java
@@ -15,7 +15,7 @@ public class RepRequest {
 	private int price;
 	private int totalPrice;
 	private int rptNo;
-	private int docno;
+	private Long docno;
 	private String mname;
 	private String sname;
 	private int smno;
